feat(elliptic_curve): impl `PointSetOp`
If a set of points created from `tensor.from_elements`
(`%a = tensor.from_elements %p1, %p2`) were to be lowered to field, the
expected output would be a tensor of tensors (ex. `tensor<2xtensor<3x!PF>>`);
however, a tensor of tensors is not allowed. Therefore, a different "data
structure" of a point_set was created such that when
`ec.point_set.from_elements` is lowered to field, the output would be
`tensor<2x3x!PF>` or a flattened version