test(mod_arith): fix filecheck assertion
The output of `printMemrefI32` would be like:
```
Unranked Memref base@ = 0x121707440 rank = 1 offset = 0 sizes = [1] strides = [1] data =
[0]
```
where the passed data is 0.
However when FileCheck meets the assertion `// CHECK: [1]`, it misinterprets that the output correctly matches the pattern because of `sizes = [1]` or `strides = [1]`.

As `sizes` can have value from 1 to 8, it's better to avoid checking one single digit value. Or we can more precisely denote the intended pattern like below, which is applied in this commit:
```
// CHECK: data =
// CHECK-NEXT: [1]
```