fix(field): rename `PowOp` to `PowUIOp` and remove sign check of exponent
The original PowOp was designed to reject negative exponents since negative
powers require inverse operations. However, it incorrectly rejected unsigned
integer exponents with the MSB set to 1, interpreting them as negative values.

Following the MLIR's design choice of builtin operations, instead of imposing
the signedness constraints on operand types or checking the MSB of input, we
explicitly indicate the intention in the operation name that the operation
interprets the exponent input as an unsigned integer by renaming PowOp to
PowUIOp.