fix(elliptic_curve): fix usage of future result as value
`DimOp` takes in a runtime value to return its dimension. However,
the operation result is not a value yet. This was working because of
`DimOp` canonicalization patterns but is not reliable and a bad usage.

In case of dynamic result shape, we have to infer the shape from input
values but it's not possible in `BucketAcc` so I will just use result
type shape.