fix: fix `scalarByPoint` logic error
The patch only allowed for scalar & point tensors to be inputted into a linalg
function in that specific order. This became a problem for matvec in bucket
reduce where we do tensor<RxCxpoint> * tensor<Cxscalars> which is the opposite
order. Thus, this commit supports different ordering of scalar/point input
types.